Added: smooth scrolling

+++ Code/Game/Display/Camera.h
@@ -11,8 +11,8 @@ struct Camera
 	uint64_t mX;
 	uint64_t mY;
 	// sub pixel positions on (top-left) tile
-	uint64_t mSubX;
-	uint64_t mSubY;
+	float  mSubX;
+	float  mSubY;
 	// count of tiles on one camera screen
 	uint64_t mWidth;
 	uint64_t mHeight;
